requirements.txt
# Core dependencies for Resume-Collector (Python)
datasets>=2.12.0
aiohttp>=3.8.0
pandas>=2.0.0
pydantic>=1.10.0
python-dotenv>=1.0.0
diskcache>=5.6.0
spacy>=3.0.0
scikit-learn>=1.2.0
pytest>=7.0.0
openai>=1.0.0
